+++ CHANGES.md
@@ -2,8 +2,8 @@ master (in development)
 =======================
 
 - Add `flycheck-pos-tip-timeout` [GH-9]
-- Add `flycheck-pos-tip-hide-function`
 - Fall back to echo area display of errors on TTY frames [GH-11] [GH-14]
+- Remove `flycheck-pos-tip-show-function`
 
 0.1 (Nov 16, 2015)
 ==================